requirements.txt
Django>=4.1
djangorestframework
celery[redis]>=5.2
pika>=1.3.0
dnspython>=2.3.0
psycopg2-binary
redis
python-dotenv